# api/swagger.json
{
  "swagger": "2.0",
  "info": {},
  "paths": {
...
    "/categories": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Get list of categories.",
        "operationId": "getCategories",
        "responses": {
          "200": {
            "description": "List of categories.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Category"
              }
            }
          },
          "400": {
            "description": "Invalid body request.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Category"
        ],
        "summary": "Create a category.",
        "operationId": "createCategory",
        "parameters": [
          {
            "name": "Category",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateCategory"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Category created.",
            "headers": {
              "Location": {
                "type": "string",
                "description": "The path of the new category created ."
              }
            }
          },
          "400": {
            "description": "Invalid body request.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Get a category by id.",
        "operationId": "getCategory",
        "parameters": [
          {
            "type": "string",
            "description": "category ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Category details.",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          },
          "400": {
            "description": "Invalid body request.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "summary": "Delete a category.",
        "operationId": "deleteCategory",
        "parameters": [
          {
            "type": "string",
            "description": "category ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted."
          },
          "400": {
            "description": "Invalid body request.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/resources": {
      "get": {
        "tags": [
          "Resource"
        ],
        "summary": "Get all available resources filtered. By default, it gives them all.",
        "operationId": "getResources",
        "parameters": [
          {
            "description": "Match or partial match of title",
            "name": "title",
            "in": "query"
          },
          {
            "description": "Match or partial match of link",
            "name": "link",
            "in": "query"
          },
          {
            "description": "comma separated list of categories",
            "name": "categories",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of resources",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResourceResponse"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Resource"
        ],
        "summary": "Create a resource.",
        "operationId": "createResource",
        "parameters": [
          {
            "name": "Resource",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateResource"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Resource created.",
            "headers": {
              "Location": {
                "type": "string",
                "description": "The path of the new resource created."
              }
            }
          },
          "400": {
            "description": "Invalid body request.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/resources/{id}": {
      "get": {
        "tags": [
          "Resource"
        ],
        "summary": "Get a resource by its id.",
        "operationId": "getResource",
        "parameters": [
          {
            "type": "string",
            "description": "resource ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Resource detail",
            "schema": {
              "$ref": "#/definitions/ResourceResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resource"
        ],
        "summary": "Delete a resource.",
        "operationId": "deleteResource",
        "parameters": [
          {
            "type": "string",
            "description": "resource ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Resource deleted"
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Resource"
        ],
        "summary": "Update an existing resource.",
        "operationId": "updateResource",
        "parameters": [
          {
            "name": "Resource",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateResource"
            }
          },
          {
            "type": "string",
            "description": "resource ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Resource updated"
          },
          "400": {
            "description": "Invalid body request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "categoryId": {
          "description": "CategoryId of category",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ID",
          "example": 2
        },
        "parentId": {
          "description": "ParentId of parent category, making it a subcategory",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ParentId",
          "example": 1
        },
        "title": {
          "description": "Title of the category",
          "type": "string",
          "x-go-name": "Title",
          "example": "backend"
        }
      },
      "x-go-package": "miniWiki/internal/domain/category/model"
    },
...
    "CreateCategory": {
      "description": "CreateCategory Category creation request",
      "type": "object",
      "properties": {
        "ID": {
          "type": "integer",
          "format": "int64"
        },
        "parentId": {
          "description": "ID of parent category, making it a subcategory",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ParentId",
          "example": 1
        },
        "title": {
          "description": "Title of the category",
          "type": "string",
          "x-go-name": "Title",
          "example": "backend"
        }
      },
      "x-go-package": "miniWiki/internal/domain/category/model"
    },
    "CreateResource": {
      "type": "object",
      "required": [
        "title",
        "link"
      ],
      "properties": {
        "categoryId": {
          "description": "Id of the category that the resource is a part of",
          "type": "integer",
          "format": "int64",
          "x-go-name": "CategoryId",
          "example": 1
        },
        "description": {
          "description": "Description of resource",
          "type": "string",
          "x-go-name": "Description",
          "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry."
        },
        "link": {
          "description": "Link of resource",
          "type": "string",
          "x-go-name": "Link",
          "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry."
        },
        "state": {
          "description": "State of resource, can be either PUBLIC or PRIVATE",
          "type": "string",
          "x-go-name": "State",
          "example": "PUBLIC"
        },
        "title": {
          "description": "Title of resource",
          "type": "string",
          "x-go-name": "Title",
          "example": "Lorem ipsum"
        }
      },
      "x-go-package": "miniWiki/internal/domain/resource/model"
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "detail": {
          "description": "Detail, more detailed message about the problem",
          "type": "string",
          "x-go-name": "Detail"
        },
        "message": {
          "description": "Message, usually for the user",
          "type": "string",
          "x-go-name": "Message"
        },
        "status": {
          "description": "Status code of response",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "miniWiki/pkg/transport"
    },
...
    "ResourceResponse": {
      "description": "ResourceResponse Resource information",
      "type": "object",
      "properties": {
        "authorId": {
          "description": "AuthorId ID of resource's author",
          "type": "integer",
          "format": "int64",
          "x-go-name": "AuthorId",
          "example": 1
        },
        "categoryId": {
          "description": "CategoryId ID of the category that the resource is a part of",
          "type": "integer",
          "format": "int64",
          "x-go-name": "CategoryId",
          "example": 1
        },
        "description": {
          "description": "Description of resource",
          "type": "string",
          "x-go-name": "Description",
          "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry."
        },
        "link": {
          "description": "Link of resource",
          "type": "string",
          "x-go-name": "Link",
          "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry."
        },
        "resourceId": {
          "description": "Id of resource",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ResourceId",
          "example": 1
        },
        "state": {
          "description": "State of resource, can be either PUBLIC or PRIVATE",
          "type": "string",
          "x-go-name": "State",
          "example": "PUBLIC"
        },
        "title": {
          "description": "Title of resource",
          "type": "string",
          "x-go-name": "Title",
          "example": "Lorem ipsum"
        }
      },
      "x-go-package": "miniWiki/internal/domain/resource/model"
    },
...
    "UpdateResource": {
      "type": "object",
      "properties": {
        "categoryId": {
          "description": "Id of the category that the resource is a part of",
          "type": "integer",
          "format": "int64",
          "x-go-name": "CategoryId",
          "example": 1
        },
        "description": {
          "description": "Description of resource",
          "type": "string",
          "x-go-name": "Description",
          "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry."
        },
        "link": {
          "description": "Link of resource",
          "type": "string",
          "x-go-name": "Link",
          "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry."
        },
        "state": {
          "description": "State of resource, can be either PUBLIC or PRIVATE",
          "type": "string",
          "x-go-name": "State",
          "example": "PUBLIC"
        },
        "title": {
          "description": "Title of resource",
          "type": "string",
          "x-go-name": "Title",
          "example": "Lorem ipsum"
        }
      },
      "x-go-package": "miniWiki/internal/domain/resource/model"
    }
  }
}